+++ services/conversation/index.js
@@ -1,0 +1,11 @@
+const { buildConversationState, initialConversationState } = require('./conversation-state');
+const { applyDialoguePolicy, sanitizeConsultantText } = require('./dialogue-policy');
+const { PERSONA } = require('./persona');
+
+module.exports = {
+  PERSONA,
+  applyDialoguePolicy,
+  buildConversationState,
+  initialConversationState,
+  sanitizeConsultantText,
+};